Move tableHead to a new component fo a CourtCaseList
